added templates for clients, servers, and subscribers
