--- Dockerfile.osiris_science
@@ -1,27 +1,0 @@
-FROM ubuntu:20.04
-
-ENV LANG C.UTF-8
-ENV LC_ALL C.UTF-8
-
-RUN apt update && apt install -y curl gnupg2 lsb-release
-RUN curl -sSL https://raw.githubusercontent.com/ros/rosdistro/master/ros.key -o /usr/share/keyrings/ros-archive-keyring.gpg
-RUN echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/ros-archive-keyring.gpg] http://packages.ros.org/ros2/ubuntu $(. /etc/os-release && echo $UBUNTU_CODENAME) main" | tee /etc/apt/sources.list.d/ros2.list > /dev/null
-
-RUN apt update
-RUN apt upgrade
-RUN DEBIAN_FRONTEND=noninteractive apt install -y ros-foxy-desktop ros-dev-tools
-
-WORKDIR /root
-RUN git clone https://github.com/SC-Robotics-2021/osiris_science.git -b docker
-WORKDIR /root/osiris_science
-
-RUN apt-get install python3-rosdep -y
-RUN rosdep init
-RUN rosdep update
-RUN rosdep install -i --from-path src --rosdistro foxy -y
-RUN apt install python3-colcon-common-extensions -y
-
-RUN colcon build --symlink-install
-
-ENTRYPOINT ["/root/osiris_science/entrypoint.sh"]
-CMD ["bash"]
