+++ entrypoint.sh
@@ -1,4 +1,2 @@
 #!/bin/bash
 
-source /opt/ros/foxy/desktop
-source ./install/local_setup.bash
