+++ src/science_clients/science_clients/uv_light_client.py
@@ -1,7 +1,7 @@
-import template_clients.service_client as service_client
+from template_clients import *
 
 def main():
-    service_client.boot(subsystem='science', component_name='uv_light')
+    boot(ServiceClient(subsystem='science', component_name='uv_light'))
     
 if __name__ == '__main__':
     main()
