+++ src/science_clients/science_clients/vacuum_client.py
@@ -1,7 +1,7 @@
-import template_clients.service_client as service_client
+from template_clients import *
 
 def main():
-    service_client.boot(subsystem='science', component_name='vacuum')
+    boot(ServiceClient(subsystem='science', component_name='vacuum'))
     
 if __name__ == '__main__':
     main()
