+++ src/science_servers/science_servers/microscope_server.py
@@ -1,9 +1,8 @@
-from server_template import *
+from template_servers import *
 
 def main():
     camera_settings = {'device_num': 2, 'fps': 30, 'resolution': (320, 240)}
-    camera_server = CameraServer(subsystem='science', component_name='microscope', camera_settings=camera_settings)
-    boot(camera_server)
+    boot(CameraServer(subsystem='science', component_name='microscope', camera_settings=camera_settings))
 
 if __name__ == '__main__':
     main()
