+++ src/science_subscribers/science_subscribers/uv_camera_sub.py
@@ -1,75 +1,8 @@
-import rclpy
-from rclpy.node import Node
-from rclpy.callback_groups import ReentrantCallbackGroup
-from rclpy.qos import qos_profile_sensor_data
-from sensor_msgs.msg import CompressedImage
-import cv2
-from cv_bridge import CvBridge
-import os
-import sys
-from datetime import datetime
-import numpy as np
+from template_subscribers import *
 
 
-class UVCameraSub(Node):
-    """
-    Subclass of ROS2 Node class. Receives and displays video feed from camera.
-    """
-
-    def __init__(self):
-        """
-        Class constructor to set up the node
-        """
-        # initiate the Node class's constructor and give it a name
-        super().__init__('uv_camera_sub')
-        self.callback_group = ReentrantCallbackGroup()
-        self.subscription = self.create_subscription(CompressedImage, '/osiris/science/uv_camera/images',
-                                                     self.receive_frame, qos_profile=qos_profile_sensor_data,
-                                                     callback_group=self.callback_group)
-        self.bridge = CvBridge()
-        frame_width = 320
-        frame_height = 240
-        self.frame = np.zeros([frame_width, frame_height, 3], dtype=np.uint8)
-        self.media_path = os.path.join(os.path.dirname(os.path.realpath(sys.argv[0]))   , 'science_recordings', 'uv_camera')
-        try:
-            os.makedirs(self.media_path, exist_ok=True)
-        except OSError as e:
-            self.get_logger().info(f'{e}')
-        now = datetime.now().strftime('%m-%d-%Y_%H:%M:%S')
-        self.out = cv2.VideoWriter(os.path.join(self.media_path, 'UV_Camera_'+now+'.mp4'),
-                                   cv2.VideoWriter_fourcc(*'mp4v'), 30, (frame_width, frame_height))
-        self.snapshot = False
-
-    def receive_frame(self, frame):
-        """
-        ROS2 subscriber callback. Receives frame via ROS2 topic. Saves image to mp4 if requested
-        """
-        self.frame = self.bridge.compressed_imgmsg_to_cv2(frame)
-        self.out.write(self.frame)
-        self.get_logger().info('Receiving UV camera frame')
-        if self.snapshot:
-            now = datetime.now().strftime('%m-%d-%Y_%H:%M:%S')
-            cv2.imwrite(os.path.join(self.media_path, 'UV_Camera_' + now + '.jpeg'), self.frame)
-            self.snapshot = False
-
-
-def main(args=None):
-    """
-    Main function. Initializes and spins ROS2 node. Camera capture and node are ended explicitly before shutdown
-    :param args: ROS2 command line arguments
-    """
-    rclpy.init(args=args)
-    uv_camera_sub = UVCameraSub()
-    try:
-        rclpy.spin(uv_camera_sub)
-    except Exception as e:
-        uv_camera_sub.get_logger().info(f'{e})')
-    except KeyboardInterrupt:
-        print('\n')
-    finally:
-        uv_camera_sub.out.release()
-        uv_camera_sub.destroy_node()
-        rclpy.shutdown()
+def main():
+    boot(CameraSubscriber(subsystem='science', component_name='uv_camera'))
 
 
 if __name__ == '__main__':
